requirements.txt
discord.py[voice]>=2.3.2
wavelink>=3.2.0